Add files via upload
Note: For excel file download from the KenPom website, I used a google extension to load data into an excel table. Any future data should be cleaned and calculated in the same fashion as the data I use. Win probability is calculated as W/(W + L), and the W-L must be spliced for this calculation to work. 

While this data has low accuracy for predicting ranking, it is showing promising results as I choose different predictor subsets. However, it is unsuccessful at predicting championship winners due to the incredibly low percentage of winners in the ~23 years of KenPom data. Still working on a solution for this issue.

Reads out a data file with the predicted ranking and predicted championship status (1/0) for each team.